Load pulseaudio at startup
PulseAudio, as shipped in Fedora, is set up through socket activation.
This means PulseAudio is not running normally until someone tries to
connect to it for the first time. The first application that tries to
access an audio device, directly or indirectly, will have to wait just a
few seconds before PulseAudio starts.

With this 'fix', PulseAudio will be started in the background together
with other processes and will be instantly available the first time a
program requests access to it.

Full description: http://rg3.name/201511101847.html